# grpc: fix regression by freeing request bufferslice after processing unary