@PapaCharlie 2024-08-29
Yeah looking now. The tests are passing so I'm not too concerned there, but I'm surprised I missed this. This isn't restricted to unary RPCs, so it may need fixes elsewhere!