@PapaCharlie 2024-08-29
Nevermind, this is the equivalent for stream RPCs: https://github.com/grpc/grpc-go/blob/master/rpc_util.go#L923

Nice catch! I just have a tiny nit, for consistency with the other places `defer Free` is called