@dfawley 2024-08-29
Let's do a patch release of 1.66 with this fix when it's ready.