@sluongng 2024-09-06
Will there be a v1.66.1 release with this fix included?